+++ src/sub_dirs.ml
@@ -1,7 +1,7 @@
 open! Stdune
 
 type 'set t =
-  { sub_dirs : 'set
+  { dirs : 'set
   ; data_only : 'set
   }
 
@@ -11,18 +11,17 @@ let default =
       | "" -> false
       | s -> s.[0] <> '.' && s.[0] <> '_')
   in
-  { sub_dirs = standard_dirs
+  { dirs = standard_dirs
   ; data_only = Predicate_lang.empty
   }
 
-let make ~sub_dirs ~data_only ~ignored_sub_dirs =
-  let sub_dirs =
-    Option.value sub_dirs ~default:default.sub_dirs in
+let make ~dirs ~data_only ~ignored_sub_dirs =
+  let dirs = Option.value dirs ~default:default.dirs in
   let data_only =
     let data_only = Option.value data_only ~default:default.data_only in
     Predicate_lang.union (data_only :: ignored_sub_dirs)
   in
-  { sub_dirs ; data_only }
+  { dirs ; data_only }
 
 let add_data_only_dirs t ~dirs =
   { t with data_only =
@@ -30,15 +29,13 @@ let add_data_only_dirs t ~dirs =
                [t.data_only; (Predicate_lang.of_string_set dirs)] }
 
 let eval t ~dirs =
-  let sub_dirs =
-    Predicate_lang.filter t.sub_dirs ~standard:default.sub_dirs dirs
-  in
+  let dirs = Predicate_lang.filter t.dirs ~standard:default.dirs dirs in
   let data_only =
-    Predicate_lang.filter t.data_only ~standard:default.data_only sub_dirs
+    Predicate_lang.filter t.data_only ~standard:default.data_only dirs
     |> String.Set.of_list
   in
-  let sub_dirs = String.Set.of_list sub_dirs in
-  { sub_dirs
+  let dirs = String.Set.of_list dirs in
+  { dirs
   ; data_only
   }
 
@@ -47,9 +44,7 @@ module Status = struct
 end
 
 let status t ~dir =
-  match String.Set.mem t.sub_dirs dir
-      , String.Set.mem t.data_only dir
-  with
+  match String.Set.mem t.dirs dir, String.Set.mem t.data_only dir with
   | true, false  -> Status.Normal
   | true, true   -> Data_only
   | false, false -> Ignored
@@ -89,25 +84,24 @@ let decode =
     Predicate_lang.decode
   in
   let decode =
-    let%map ignored_sub_dirs =
-      multi_field "ignored_subdirs" ignored_sub_dirs
+    let%map dirs = field_o "dirs" (located plang)
     and data_only = field_o "data_only_dirs" (located plang)
-    and sub_dirs = field_o "subdirs" (located plang)
+    and ignored_sub_dirs = multi_field "ignored_subdirs" ignored_sub_dirs
     and rest = leftover_fields
     in
-    match data_only, sub_dirs, ignored_sub_dirs with
+    match data_only, dirs, ignored_sub_dirs with
     | None, Some (loc, _), _::_ ->
       Errors.fail loc
-        "Cannot have both sub_dirs and ignored_subdirs \
+        "Cannot have both dirs and ignored_subdirs \
          stanza in a dune file. "
     | Some (loc, _), None, _::_ ->
       Errors.fail loc
         "Cannot have both data_only_dirs and ignored_subdirs \
          stanza in a dune file. "
     | _ ->
-      let sub_dirs = Option.map ~f:snd sub_dirs in
+      let dirs = Option.map ~f:snd dirs in
       let data_only = Option.map ~f:snd data_only in
-      ( make ~sub_dirs ~data_only ~ignored_sub_dirs
+      ( make ~dirs ~data_only ~ignored_sub_dirs
       , rest
       )
   in